Remove unnecessary java name
