+++ src/main/java/se/iths/charity_shop/entity/StockEntity.java
@@ -1,4 +1,4 @@
-package java.se.iths.charity_shop.entity;
+package se.iths.charity_shop.entity;
 
 
 import javax.persistence.Entity;
@@ -7,7 +7,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Entity
-public class CharityEntity {
+public class StockEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
