+++ src/main/java/se/iths/charity_shop/repository/UserRepository.java
@@ -1,8 +1,8 @@
-package java.se.iths.charity_shop.repository;
+package se.iths.charity_shop.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import java.se.iths.charity_shop.entity.UserEntity;
+import se.iths.charity_shop.entity.UserEntity;
 
 public interface UserRepository extends CrudRepository {
     UserEntity findByUsername(String username);
