+++ src/test/java/se/iths/charity_shop/CharityShopApplicationTests.java
@@ -1,4 +1,4 @@
-package java.se.iths.charity_shop;
+package se.iths.charity_shop;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
